hotfix/repetition-selection
Selection wasn't working on repetitions of components
with nested components in certain circunstances.

firstDraw event bubbles, causing childComponentsLeftToDraw
to become a negative number as the number of sub-components
that drew might be higher than the number of childComponents,
and so _iterationForElement not containing all of the iterations
it should contain.

By checking that the target is the currentTarget we fix the issue.