Fix non-English Localizer loading tests
There were a couple of tests trying to initWithLocale before
require was set, and so the appropiate manifest for the tests.
As initWithLocale calls loadMessages, it was causing a
unhandled promise rejection. By setting the require before
the initWithLocale I fix the tests.

However I smell some bad design in the Localizer, as
initWithLocale should probably not be an initializer but more
of a setLocale that returns the loadMessages promise. I will
leave this idea for future discussion.