Remove data spec tests testing native features
There are some old tests testing features of the now
not used Date.parse shim. As the shim is not being in
use right now, the tests are testing native features.
Depending on the browser the support for ISO 8601
date parsing might be more or less complete, causing
some of these tests to fail. I have just xit-ed them in
case we want to start using again the shim in the future