Fix Date test
It seems that after a switch from Date.parse shim to
native Date.parse this concrete test wasn't updated
properly. The shim returns a Date object, while the
native one returns a number of miliseconds. I have
left some comments in case in the future the shim
is started to be used again for some reason.