Fix montage serializer tests
There were a couple of tests failing. The reason was that
they use a spy in the function fakeGetSerializablePropertyNames
whose result now is memoized (I assume it wasn't when the
tests/spy was created), so the result was not the expected.

After fixing it I tried to understand the reason for those tests,
and I couldn't. So I went back in history to understand them
and they have been silly tests for the last 12 years. Maybe they
made sense before, but the git history only reaches until there.

So given that they don't make sense and that other tests
cover the testing of multiple references to the same object
I have decided to kill them with fire.