Increase Flow frustum culling numerical stability
Numerical precission issues were affecting Flow for certain
paths, and this was causing a couple of tests to fail.

The code is still not very numerically stable as cubicRealRoots
is using the algebraic formula that is limited. In the future we
should consider implementing a different algorithm to find the
plane/bezier intersections such as bezier recursive subdivision
and intersection with bounding boxes that is way more robust.