Update data-service test to new behaviour
Data Service behaviour regarding mainService
was updated but not the matching test, causing
it to fail. I have updated the test and added a
new one to match the new behaviour.

I have also removed a return in a Promise to
avoid the deprecation in Jasmine of using
promise return and done at the same time.