Update data-service test to latest behaviour
One of the data-service tests was failing because
it was expecting the old mainService behaviour
were the first constructed service would be the
one assigned to mainService.

I have updated the test to the new way it works so
I manually set the service to the application's
mainService and this way the test is passing again.