Fix failing HttpService test
The code in Data Service was assuming that any
ObjectDescriptor with a module has the
loadObjectFromModule method from the
ModuleObjectDescriptor. However the failing tests
prove that this is not necessarily true.

It is debatable if the test is correct or not, and if
non ModuleObjectDescriptors should have modules
or not, but I feel that the checks I'm adding can not
hurt, so I feel this is a reasonable fix.