Fix regression in module object descriptor spec
There was a regression that I probably attribute
to .gitignore ignoring all node_modules. I have
updated it so it only ignores the root node_modules
so the node_modules related to tests are tracked.

Then I have a restablished a file that was deleted
and now 2 failing tests are now passing.